:sparkles: Daily report now can save value to db, decrease request number

+++ app/model/model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, Enum, Boolean, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, Date, Enum, Boolean, ForeignKey, DateTime, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from datetime import datetime, timezone
@@ -39,7 +39,7 @@ class DailyMargin(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     date = Column(Date, default=datetime.today, nullable=False)
     margin_ticket = Column(Integer, nullable=False)
-    margin_amount = Column(Integer, nullable=False)
+    margin_amount = Column(Float, nullable=False)
 
 # Daily Future table
 class DailyFuture(Base):
